mini-project-1/requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.1.0
langgraph>=0.4.3
langgraph-checkpoint>=2.0.25
openai>=1.0.0
python-dotenv>=1.0.0
pdfminer.six==20200517
numpy>=1.24.0
urllib3>=1.26.0
langsmith>=0.1.0
rouge-score==0.1.2 